mobile error

+++ migrations/20230506094305-create_table_visitors_info.js
@@ -19,7 +19,7 @@ module.exports = {
           type: Sequelize.STRING(50)
       },
       mobile:{
-          type: Sequelize.INTEGER(10)
+          type: Sequelize.INTEGER(50)
       },
       subject:{
           type: Sequelize.STRING(50)
